Stop using deprecated Corvus.SpecFlow.Extensions (#201)
* Now using Corvus.Testing.SpecFlow
* Renamed Corvus.SpecFlow.Extensions.CosmosClient to Corvus.Testing.CosmosDb.SpecFlow to match
* Restructured namespaces inside Corvus.Testing.CosmosDb.SpecFlow - the existing structure was already a bit odd, but now also looked inconsistent with Corvus.Testing.SpecFlow

Note that although the namespace restructuring is a breaking change, we've also renamed the component, so there's no need for a major version change. Instead, we will need to mark Corvus.SpecFlow.Extensions.CosmosClient as deprecated once we release a non-preview version of its replacement, Corvus.Testing.CosmosDb.SpecFlow. (This is in line with deprecation of Corvus.SpecFlow.Extensions in favour of Corvus.Testing.SpecFlow.)