Adjust batch limiting to ignore empty batches
We've started seeing the Cosmos Client return empty batches. (It's not clear if this is due to the change from `Microsoft.Azure.Cosmos` v3.30 to v3.31 or if it's something that has changed on the server side.)

The tests are set up not to expect empty batches. E.g., there are a couple that make a query that should return 2 results, they ask for a maximum batch size of 1, a maximum batch count of 2, and expect to get all the results.

That doesn't seem unreasonable, but we're now seeing the Cosmos client produce 3 batches, one of which is empty, in completing this request! So when it comes to counting the number of batches, we now disregard empty batches.